Use managed TsString and friends
* Use C# TsStrFactory and TsPropsFactory in TsStringUtils
* Remove calls to Marshal.ReleaseComObject() on TsStrings
* Intern TsTextProps and empty TsStrings
* Fix issue with marshaling in NfdAndFixOffsets() calls
* Properly handle getting TsTextProps properties that don't return
S_FALSE when the property isn't set
* Fix bug in TsString normalization
* Only remove a string property when TsTextProps.SetStrPropValue()
is called with a null value and not with an empty value
* Fix endianness of object data stored as string properties in
TsTextProps
* Ensure that C++ COM objects are released before any C# objects that
they hold references to are garbage-collected  (deals with GC bug in
Mono)

+++ Src/Common/CoreImpl/TsPropsSerializer.cs
@@ -1500,7 +1500,7 @@ namespace SIL.CoreImpl
 		/// ------------------------------------------------------------------------------------
 		private static void AddObjDataToBldr(ITsPropsBldr propsBldr, XAttribute attr, FwObjDataTypes type)
 		{
-			byte[] objData = TsStringUtils.GetObjData(new Guid(attr.Value), (byte)type);
+			byte[] objData = TsStringUtils.GetObjData(new Guid(attr.Value), type);
 			propsBldr.SetStrPropValueRgch((int)FwTextPropType.ktptObjData, objData, objData.Length);
 		}
 
@@ -1514,7 +1514,7 @@ namespace SIL.CoreImpl
 		/// ------------------------------------------------------------------------------------
 		private static void AddObjDataToBldr(ITsIncStrBldr strBldr, XAttribute attr, FwObjDataTypes type)
 		{
-			byte[] objData = TsStringUtils.GetObjData(new Guid(attr.Value), (byte)type);
+			byte[] objData = TsStringUtils.GetObjData(new Guid(attr.Value), type);
 			strBldr.SetStrPropValueRgch((int)FwTextPropType.ktptObjData, objData, objData.Length);
 		}
